Move LockedLocation enum outside of constTurret
