+++ src/main/java/frc/robot/subsystems/Shooter.java
@@ -22,7 +22,7 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants.constShooter;
-import frc.robot.Constants.constTurret.LockedLocation;
+import frc.robot.Constants.LockedLocation;
 import frc.robot.RobotMap.mapShooter;
 import frc.robot.RobotPreferences.prefShooter;
